feat(installer): Add advanced commit handling
The `commit` value of the plugin_spec can now be either a `string` *or*
a `table`. The table's value lets you define both the commit `hash` and
the `date` of the hash. This enables the use of `--shallow-since` for
clones. Instead of setting a depth of `999999` it will shallow clone
from the date provided.

The date of a commit can be fould using a command like:

```
git show -s --format="%cI" HEAD
```